Move development-related decorators to specific module
