+++ doc/releases/v0.11.0.txt
@@ -2,6 +2,8 @@
 v0.11.0 (Unreleased)
 --------------------
 
+- Enforced keyword-only arguments for most parameters of most functions and classes.
+
 - Added an explicit warning in :func:`swarmplot` when more than 2% of the points are overlap in the "gutters" of the swarm.
 
 - Added the ``axes_dict`` attribute to :class:`FacetGrid` for named access to the component axes.
