+++ seaborn/distributions.py
@@ -14,8 +14,9 @@ try:
 except ImportError:
     _has_statsmodels = False
 
-from .utils import iqr, _kde_support, remove_na, _deprecate_positional_args
+from .utils import iqr, _kde_support, remove_na
 from .palettes import color_palette, light_palette, dark_palette, blend_palette
+from ._decorators import _deprecate_positional_args
 
 
 __all__ = ["distplot", "kdeplot", "rugplot"]
